Split metadata table into separate modules (#872)
